Refactor services out of Main component into Services component
