+++ react-app/src/Main.scss
@@ -6,56 +6,3 @@ main {
 main h1 {
   margin: 36px 0 21px 0;
 }
-
-div.div--search {
-  background-color: #F2F2F2;
-  margin: 0 0 20px 0;
-  padding: 30px;
-}
-
-div.div--search h2 {
-  margin: 0 0 18px 0;
-}
-
-div.div--search form {
-  background-color: white;
-  border: 1px solid #D1D1D1;
-  box-sizing: border-box;
-  display: inline-block;
-  height: 46px;
-  width: 100%;
-}
-
-div.div--search form input {
-  border: 0;
-  box-sizing: border-box;
-  display: inline-block;
-  font-family: "BCSans", "Noto Sans", Verdana, Arial, sans-serif;
-  font-size: 24px;
-  height: 44px;
-  width: calc(100% - 44px);
-  vertical-align: middle;
-}
-
-div.div--search form button {
-  appearance: button;
-  -moz-appearance: button;
-  -webkit-appearance: button;
-  background: url("data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIxOC40NzQiIGhlaWdodD0iMTguNDc3IiB2aWV3Qm94PSIwIDAgMTguNDc0IDE4LjQ3NyI+CiAgPHBhdGggaWQ9InNlYXJjaC1taW51cyIgZD0iTTE4LjIyMiwxNy4yLDE3LjIsMTguMjIyYS44NjIuODYyLDAsMCwxLTEuMjIzLDBsLTMuNi0zLjZhLjg2NS44NjUsMCwwLDEtLjI1My0uNjEzdi0uNTg4YTcuNSw3LjUsMCwxLDEsMS4zLTEuM2guNTg4YS44NjUuODY1LDAsMCwxLC42MTMuMjUzbDMuNiwzLjZBLjg3My44NzMsMCwwLDEsMTguMjIyLDE3LjJabS01LjgwOS05LjdhNC45MDcsNC45MDcsMCwxLDAtNC45MDcsNC45MDdBNC45LDQuOSwwLDAsMCwxMi40MTMsNy41MDVaIi8+Cjwvc3ZnPgo=") no-repeat;
-  background-position: center;
-  border: 0;
-  box-sizing: border-box;
-  display: inline-block;
-  height: 44px;
-  padding: 0;
-  width: 44px;
-  vertical-align: middle;
-}
-
-div.div--service-highlight-container {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  column-gap: 20px;
-  row-gap: 20px;
-  margin: 0 0 20px 0;
-}
